% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{nba_pgstats}
\alias{nba_pgstats}
\title{NBA Player PG Stats}
\usage{
nba_pgstats(player)
}
\arguments{
\item{player}{A string containing the player's full name}
}
\value{
Returns a data frame containing the career per game stats by season
}
\description{
Scrape the career per game statistics for a given player
}
