% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{player_data}
\alias{player_data}
\title{Player Per Game Stats}
\usage{
player_data(player, league, advanced = F)
}
\arguments{
\item{player}{A string containing the player's full name}

\item{league}{A string containing the league to be searched. One of: 'NFL', 'NBA', 'MLB', 'NHL', 'CBB'.}

\item{advanced}{Default is false. Make true to retrieve advanced NBA, CBB, and MLB stats.}
}
\value{
Returns a data frame containing the career per game stats by season. Descriptions of statistics
    can be found by searching the following sites:
    \itemize{
       \item www.basketball-reference.com
       \item www.pro-football-reference.com
       \item www.hockey-reference.com
       \item www.baseball-reference.com
       \item www.sports-reference.com/cbb
    }
}
\description{
Scrape the career per game statistics for a given player
}
\examples{
\dontrun{
   df <- player_data("Kobe Bryant", "NBA")
   df <- player_data("Mike Trout", "MLB", advanced = TRUE)
}
}
