% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season_stats.R
\name{season_stats}
\alias{season_stats}
\title{Scrape Season Stats}
\usage{
season_stats(year, league)
}
\arguments{
\item{league}{A string containing the league to search. One of: 'NFL', 'NBA', 'NHL', 'MLB'}

\item{team}{A year to get statistics for. For season that occur across two years, use the first.}

\item{defensive}{Whether to return defensive stats or offensive. Only applys for NBA and MLB teams}
}
\value{
A data frame containing seasonal data for the league and year specified.
}
\description{
Scrape the seasonal statistics for a specific league and year
}
