% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfl.R
\name{team_stats}
\alias{team_stats}
\title{Scrape Team Season Stats}
\usage{
team_stats(team, league, years)
}
\arguments{
\item{team}{A string containing the full name of the team}

\item{league}{A string containing the league to search. One of: 'NFL', 'NBA', 'NHL', 'MLB'}
}
\description{
Scrape the stats for a specific range of seasons for a team
}
