% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{player_stats}
\alias{player_stats}
\title{Player PG Stats}
\usage{
player_stats(player, league)
}
\arguments{
\item{player}{A string containing the player's full name}

\item{league}{A string containing the league to be searched. One of: 'NFL', 'NBA', 'MLB', 'NHL'.}
}
\value{
Returns a data frame containing the career per game stats by season.
}
\description{
Scrape the career per game statistics for a given player
}
