% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team.R
\name{team_data}
\alias{team_data}
\title{Scrape Team Season Stats}
\usage{
team_data(team, league, defensive = F)
}
\arguments{
\item{team}{A string containing the full name of the team}

\item{league}{A string containing the league to search. One of: 'NFL', 'NBA', 'NHL', 'MLB'}

\item{defensive}{Whether to return defensive stats or offensive. Only applies for NBA and MLB teams}
}
\value{
A data frame containing seasonal data for the team. Both offensive
    and defensive return the same set of advanced statistics for the
    season and then the rest are their respective statistics.
}
\description{
Scrape the seasonal statistics for a specific team.
}
